+++ sereto/tui/finding.py
@@ -20,7 +20,7 @@ from sereto.enums import Risk
 from sereto.exceptions import SeretoValueError
 from sereto.models.finding import FindingTemplateFrontmatterModel, VarsMetadataModel
 from sereto.project import Project
-from sereto.tui.widgets.input import InputWithLabel, ListInput, SelectWithLabel
+from sereto.tui.widgets.input import InputWithLabel, ListWidget, SelectWithLabel
 
 
 @dataclass
@@ -108,10 +108,43 @@ class AddFindingScreen(ModalScreen[None]):
             for var in self.finding.variables:
                 yield Static(f"[b]{var.name}:[/b] {escape(var.type_annotation)}\n  {var.description}", classes="pl-1")
                 if var.is_list:
-                    yield ListInput(id=f"var-{var.name}")
+                    match var.type:
+                        case "boolean":
+                            yield ListWidget(
+                                widget_factory=lambda var=var: Select(  # type: ignore[misc]
+                                    options=[
+                                        ("True", True),
+                                        ("False", False),
+                                    ],
+                                    allow_blank=not var.required,
+                                ),
+                                id=f"var-{var.name}",
+                            )
+                        case "integer":
+                            yield ListWidget(
+                                widget_factory=lambda: Input(type="integer", classes="m-1"),
+                                id=f"var-{var.name}",
+                            )
+                        case _:
+                            yield ListWidget(
+                                widget_factory=lambda: Input(classes="m-1"),
+                                id=f"var-{var.name}",
+                            )
                 else:
-                    yield Input(id=f"var-{var.name}", classes="m-1")
-
+                    match var.type:
+                        case "boolean":
+                            yield Select(
+                                options=[
+                                    ("True", True),
+                                    ("False", False),
+                                ],
+                                allow_blank=not var.required,
+                                id=f"var-{var.name}",
+                            )
+                        case "integer":
+                            yield Input(id=f"var-{var.name}", type="integer", classes="m-1")
+                        case _:
+                            yield Input(id=f"var-{var.name}", classes="m-1")
                 yield Rule()
 
             self.btn_save_finding = Button.success("Save", id="save-finding", classes="m-1")
@@ -135,23 +168,62 @@ class AddFindingScreen(ModalScreen[None]):
 
         for var in self.finding.variables:
             if var.is_list:
-                all_inputs = self.query_one(f"#var-{var.name}", ListInput).query(Input).results()
-                input_values = [input.value.strip() for input in all_inputs if len(input.value.strip()) > 0]
-                if var.required and len(input_values) == 0:
+                widgets = list(self.query_one(f"#var-{var.name}", ListWidget).query(".widget").results())
+
+                match var.type:
+                    case "boolean":
+                        # get values, filter out NoSelection
+                        values = [
+                            w.value for w in widgets if isinstance(w, Select) and not isinstance(w.value, NoSelection)
+                        ]
+                    case "integer":
+                        values_str = [
+                            w.value.strip() for w in widgets if isinstance(w, Input) and len(w.value.strip()) > 0
+                        ]
+                        try:
+                            values = [int(v) for v in values_str]
+                        except ValueError:
+                            raise SeretoValueError(f"variable '{var.name}' must be an integer") from None
+                    case _:
+                        values = [
+                            w.value.strip() for w in widgets if isinstance(w, Input) and len(w.value.strip()) > 0
+                        ]
+
+                if var.required and len(values) == 0:
                     raise SeretoValueError(f"variable '{var.name}' is required")
-                else:
-                    # always set list variables, even if empty
-                    variables[var.name] = input_values
+                elif len(values) == 0:
+                    # don't set the variable if not required and empty
+                    continue
+                # always set list variables, even if empty
+                variables[var.name] = values
+                continue
             else:
-                value = self.query_one(f"#var-{var.name}", Input).value.strip()
-                if len(value) == 0:
-                    if var.required:
-                        raise SeretoValueError(f"variable '{var.name}' is required")
-                    else:
-                        # don't set the variable if not required and empty
-                        continue
-                else:
-                    variables[var.name] = value
+                match var.type:
+                    case "boolean":
+                        value_select: Select[bool] = self.query_one(f"#var-{var.name}", Select)
+                        value: int | str | None = (
+                            value_select.value if not isinstance(value_select.value, NoSelection) else None
+                        )
+                    case "integer":
+                        value_str = self.query_one(f"#var-{var.name}", Input).value.strip()
+                        if len(value_str) == 0:
+                            if var.required:
+                                raise SeretoValueError(f"variable '{var.name}' is required")
+                            else:
+                                continue
+                        try:
+                            value = int(value_str)
+                        except ValueError:
+                            raise SeretoValueError(f"variable '{var.name}' must be an integer") from None
+                    case _:
+                        value = self.query_one(f"#var-{var.name}", Input).value.strip()
+                        if len(value) == 0:
+                            if var.required:
+                                raise SeretoValueError(f"variable '{var.name}' is required")
+                            else:
+                                # don't set the variable if not required and empty
+                                continue
+                variables[var.name] = value
 
         return variables