+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.2] - 2025-06-26
+
 ### Added
 
 - Add `sereto oxipng` command to optimize PNG images in the report.
@@ -18,7 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
-- Properly handle var types (boolean/integer,string) in `sereto findings add` TUI.
+- Properly handle var types (boolean/integer/string) in `sereto findings add` TUI.
 
 ### Dependencies
 
@@ -439,7 +441,8 @@ We registered a dummy package to PyPI to test the publishing. Therefore a versio
 Initial version
 
 
-[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.3.1...HEAD
+[unreleased]: https://github.com/s3r3t0/sereto/compare/v0.3.2...HEAD
+[0.3.2]: https://github.com/s3r3t0/sereto/compare/v0.3.1...v0.3.2
 [0.3.1]: https://github.com/s3r3t0/sereto/compare/v0.3.0...v0.3.1
 [0.3.0]: https://github.com/s3r3t0/sereto/compare/v0.2.9...v0.3.0
 [0.2.9]: https://github.com/s3r3t0/sereto/compare/v0.2.8...v0.2.9